remove unnecessary type casts

+++ contracts/bridge/src/contract.rs
@@ -239,7 +239,7 @@ impl BridgeContract {
     pub fn get_transaction_cost(env: Env, chain_id: u32) -> Result<u128, Error> {
         bump_instance(&env);
 
-        get_transaction_cost(&env, chain_id as u8)
+        get_transaction_cost(&env, chain_id)
     }
 
     pub fn get_gas_usage(env: Env, chain_id: u32) -> Result<u128, Error> {
